Refactoring to move the resolution in the AbstractContext and just keeping the replacement in ContextUtils
